car_classifier/models/car_classifier/requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy
Pillow
sklearn